all controllers use proper response entities

+++ argentum-api/src/main/java/net/devium/argentum/rest/model/OrderRequest.java
@@ -3,18 +3,18 @@ package net.devium.argentum.rest.model;
 import java.util.List;
 
 public class OrderRequest {
-    private int guestId;
+    private long guestId;
     private List<OrderItemRequest> items;
 
     public OrderRequest() {
     }
 
-    public OrderRequest(int guestId, List<OrderItemRequest> items) {
+    public OrderRequest(long guestId, List<OrderItemRequest> items) {
         this.guestId = guestId;
         this.items = items;
     }
 
-    public int getGuestId() {
+    public long getGuestId() {
         return guestId;
     }
