+++ argentum-api/src/main/java/net/devium/argentum/rest/model/ProductRangeResponseMeta.java
@@ -1,18 +1,18 @@
 package net.devium.argentum.rest.model;
 
 public class ProductRangeResponseMeta {
-    private int id;
+    private long id;
     private String name;
 
     public ProductRangeResponseMeta() {
     }
 
-    public ProductRangeResponseMeta(int id, String name) {
+    public ProductRangeResponseMeta(long id, String name) {
         this.id = id;
         this.name = name;
     }
 
-    public int getId() {
+    public long getId() {
         return id;
     }
